testing/20140327.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20140327</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<script src="/irc-logs/s.js"></script>
<h1 id="w3c/testing.20140327.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2014-03-27 / <a href="#end">end</a></h1>
<p>Options: 
<ol id="lines">
 <li id="l-1"><a href="#l-1">#</a> Session Start: Thu Mar 27 00:00:00 2014</li>
 <li id="l-2"><a href="#l-2">#</a> Session Ident: #testing</li>
 <li id="l-3"><a href="#l-3">#</a> [00:01] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-4"><a href="#l-4">#</a> [00:03] * Quits: lukeis (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-5"><a href="#l-5">#</a> [00:12] * Quits: lmclister (~lmclister@public.cloak) (&quot;&quot;)</li>
 <li id="l-6"><a href="#l-6">#</a> [00:46] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-7"><a href="#l-7">#</a> [00:53] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-8"><a href="#l-8">#</a> [01:08] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-9"><a href="#l-9">#</a> [01:39] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-10"><a href="#l-10">#</a> [01:40] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-11"><a href="#l-11">#</a> [01:46] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-12"><a href="#l-12">#</a> [01:47] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-13"><a href="#l-13">#</a> [01:54] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-14" _a="heycam"><a href="#l-14">#</a> [02:12] * heycam is now known as heycam|away</li>
 <li id="l-15"><a href="#l-15">#</a> [02:12] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-16"><a href="#l-16">#</a> [02:45] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-17"><a href="#l-17">#</a> [02:47] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-18"><a href="#l-18">#</a> [02:54] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-19"><a href="#l-19">#</a> [03:09] * Joins: yankhates_cb (~yankhatescb@public.cloak)</li>
 <li id="l-20"><a href="#l-20">#</a> [03:47] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-21"><a href="#l-21">#</a> [03:48] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-22"><a href="#l-22">#</a> [03:55] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-23"><a href="#l-23">#</a> [04:20] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-24"><a href="#l-24">#</a> [04:21] * Quits: glenn (~gadams@public.cloak) (&quot;Leaving...&quot;)</li>
 <li id="l-25"><a href="#l-25">#</a> [04:36] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-26"><a href="#l-26">#</a> [04:37] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-27"><a href="#l-27">#</a> [04:39] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-28"><a href="#l-28">#</a> [04:42] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-29"><a href="#l-29">#</a> [04:42] * Quits: yankhates_cb (~yankhatescb@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-30"><a href="#l-30">#</a> [04:48] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-31"><a href="#l-31">#</a> [04:49] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-32"><a href="#l-32">#</a> [04:56] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-33" _a="heycam|away"><a href="#l-33">#</a> [05:00] * heycam|away is now known as heycam</li>
 <li id="l-34"><a href="#l-34">#</a> [05:00] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-35"><a href="#l-35">#</a> [05:49] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-36"><a href="#l-36">#</a> [05:56] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-37" _a="heycam"><a href="#l-37">#</a> [06:35] * heycam is now known as heycam|away</li>
 <li id="l-38"><a href="#l-38">#</a> [06:50] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-39"><a href="#l-39">#</a> [06:57] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-40"><a href="#l-40">#</a> [07:11] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-41"><a href="#l-41">#</a> [07:11] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-42"><a href="#l-42">#</a> [07:12] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-43"><a href="#l-43">#</a> [07:24] * Joins: ptressel (~chatzilla@public.cloak)</li>
 <li id="l-44"><a href="#l-44">#</a> [07:38] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-45"><a href="#l-45">#</a> [08:06] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-46" _a="sangwhan__"><a href="#l-46">#</a> [08:18] * sangwhan__ is now known as sangwhan</li>
 <li id="l-47" _a="sangwhan" _r="denis"><a href="#l-47">#</a> [09:19] &lt;sangwhan&gt; denis: need to update the ttwf seoul page focus points - any input?</li>
 <li id="l-48" _a="sangwhan"><a href="#l-48">#</a> [09:19] &lt;sangwhan&gt; currently just a copy and paste of the tokyo event, which is gonna end up in a boatload of duplicates of the currently stale reviews</li>
 <li id="l-49"><a href="#l-49">#</a> [09:20] * Joins: dom (dom@public.cloak)</li>
 <li id="l-50"><a href="#l-50">#</a> [09:27] * Joins: darobin (rberjon@public.cloak)</li>
 <li id="l-51" _a="@denis" _r="sangwhan"><a href="#l-51">#</a> [09:35] &lt;@denis&gt; sangwhan, you should get in touch with rebecca. I think she's the one responsible for all the ttfw events pages</li>
 <li id="l-52"><a href="#l-52">#</a> [09:35] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-53"><a href="#l-53">#</a> [09:40] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-54"><a href="#l-54">#</a> [09:40] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-55"><a href="#l-55">#</a> [09:43] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-56" _a="sangwhan" _r="denis"><a href="#l-56">#</a> [09:44] &lt;sangwhan&gt; denis: was more of a question of what is missing in web-platform-tests, actually</li>
 <li id="l-57"><a href="#l-57">#</a> [09:44] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-58" _a="Ms2ger"><a href="#l-58">#</a> [09:44] &lt;Ms2ger&gt; All the things</li>
 <li id="l-59" _a="sangwhan" _r="denis"><a href="#l-59">#</a> [09:45] &lt;sangwhan&gt; denis: i don't think we should have people sitting around and creating more stale PRs</li>
 <li id="l-60" _a="sangwhan" _r="denis"><a href="#l-60">#</a> [09:47] &lt;sangwhan&gt; denis: right now the page has the same objectives as the tokyo event (and quite likely the shenzhen event) so we're gonna end up with dupes if we keep it that way</li>
 <li id="l-61" _a="Ms2ger"><a href="#l-61">#</a> [09:48] &lt;Ms2ger&gt; We made a list of useful tests to write once</li>
 <li id="l-62" _a="Ms2ger"><a href="#l-62">#</a> [09:48] &lt;Ms2ger&gt; I don't know where it ended up</li>
 <li id="l-63" _a="@denis" _r="sangwhan"><a href="#l-63">#</a> [09:48] &lt;@denis&gt; sangwhan, I can't talk for the other specs but for HTML5, it would be nice to have more tests on www.whatwg.org/html/#browsers</li>
 <li id="l-64" _a="Ms2ger"><a href="#l-64">#</a> [09:50] &lt;Ms2ger&gt; True</li>
 <li id="l-65" _a="sangwhan" _r="denis"><a href="#l-65">#</a> [09:50] &lt;sangwhan&gt; denis: that's actually very useful information, will note</li>
 <li id="l-66" _a="Ms2ger"><a href="#l-66">#</a> [09:50] &lt;Ms2ger&gt; Not sure how useful tests for that by inexperienced contributors would be</li>
 <li id="l-67" _a="sangwhan"><a href="#l-67">#</a> [09:50] &lt;sangwhan&gt; (other wishlists are welcome)</li>
 <li id="l-68" _a="@denis" _r="jgraham"><a href="#l-68">#</a> [09:50] &lt;@denis&gt; jgraham, zcorpan, darobin, MikeSmith may have more ideas</li>
 <li id="l-69" _a="Ms2ger" _r="sangwhan"><a href="#l-69">#</a> [09:52] &lt;Ms2ger&gt; sangwhan, I think we sent that list to rhauck... Maybe she can find it</li>
 <li id="l-70" _a="sangwhan" _r="Ms2ger"><a href="#l-70">#</a> [10:10] &lt;sangwhan&gt; Ms2ger: this probably isn't the first time the inexperienced contributor discussion has popped up, it's a tricky topic</li>
 <li id="l-71" _a="sangwhan" _r="Ms2ger"><a href="#l-71">#</a> [10:11] &lt;sangwhan&gt; Ms2ger: at least there is a chance of digging up a couple good people who aren't visible atm out</li>
 <li id="l-72" _a="sangwhan"><a href="#l-72">#</a> [10:16] &lt;sangwhan&gt; how about having some people take a look at the stale reviews and see if they can address the issues raised that were never followed up on?</li>
 <li id="l-73"><a href="#l-73">#</a> [10:19] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-74" _a="sangwhan" _r="denis"><a href="#l-74">#</a> [10:26] &lt;sangwhan&gt; denis: Right now the page for Seoul mentions HTML5 Forms, CSS Basic User Interface, CSS Shapes, CSS Masking, SPARQL Query Language for RDF which I think is outdated</li>
 <li id="l-75" _a="@denis"><a href="#l-75">#</a> [10:28] &lt;@denis&gt; I think HTML5 forms is ok now. Of course we can always have more tests but it's not a priority</li>
 <li id="l-76" _a="@denis"><a href="#l-76">#</a> [10:28] &lt;@denis&gt; I can't speak for CSS and sparql</li>
 <li id="l-77" _a="sangwhan" _r="denis"><a href="#l-77">#</a> [10:34] &lt;sangwhan&gt; denis: i agree on forms, last couple events all did forms. not sure who put in sparql, but i guess that hasn't been covered</li>
 <li id="l-78"><a href="#l-78">#</a> [10:36] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-79" _a="darobin"><a href="#l-79">#</a> [10:37] &lt;darobin&gt; do we actually want to have sparql tests in WPT?</li>
 <li id="l-80" _a="darobin"><a href="#l-80">#</a> [10:37] &lt;darobin&gt; I would see that as more adapted to, say, a semantic-platform-tests repo</li>
 <li id="l-81" _a="sangwhan"><a href="#l-81">#</a> [10:38] &lt;sangwhan&gt; not sure if that's a web platform feature, looking at the specs</li>
 <li id="l-82" _a="jgraham" _r="Yeah"><a href="#l-82">#</a> [10:38] &lt;jgraham&gt; Yeah, no</li>
 <li id="l-83" _a="sangwhan"><a href="#l-83">#</a> [10:38] &lt;sangwhan&gt; and it's not a browser feature in any way from what i see</li>
 <li id="l-84" _a="sangwhan"><a href="#l-84">#</a> [10:38] &lt;sangwhan&gt; question is who put it in there i guess</li>
 <li id="l-85" _a="darobin"><a href="#l-85">#</a> [10:38] &lt;darobin&gt; right</li>
 <li id="l-86" _a="darobin" _r="well"><a href="#l-86">#</a> [10:38] &lt;darobin&gt; well, I don't care about who</li>
 <li id="l-87" _a="darobin"><a href="#l-87">#</a> [10:38] &lt;darobin&gt; and I'm happy for people to test SPARQL if they want to</li>
 <li id="l-88" _a="darobin"><a href="#l-88">#</a> [10:39] &lt;darobin&gt; it's actually a pretty good way of getting JSON out of unusable RDF data sources :)</li>
 <li id="l-89" _a="darobin"><a href="#l-89">#</a> [10:39] &lt;darobin&gt; but I reckon there should be a separate repo for that</li>
 <li id="l-90" _a="sangwhan"><a href="#l-90">#</a> [10:41] &lt;sangwhan&gt; but that's going to be a library test unless it's natively implemented in the browser (which it isn't), no?</li>
 <li id="l-91" _a="sangwhan" _r="jgraham"><a href="#l-91">#</a> [10:41] &lt;sangwhan&gt; jgraham: any way to take over critic reviews from someone else without having to reopen a review?</li>
 <li id="l-92" _a="jgraham" _r="sangwhan"><a href="#l-92">#</a> [10:42] &lt;jgraham&gt; sangwhan: You mean push more commits to the review?</li>
 <li id="l-93" _a="jgraham"><a href="#l-93">#</a> [10:42] &lt;jgraham&gt; If so the answer is &quot;no due to the suckiness of GitHub&quot;</li>
 <li id="l-94" _a="sangwhan" _r="jgraham"><a href="#l-94">#</a> [10:43] &lt;sangwhan&gt; jgraham: yes, push more commits to someone else's review</li>
 <li id="l-95" _a="sangwhan" _r="cool"><a href="#l-95">#</a> [10:43] &lt;sangwhan&gt; cool, that's what i thought too</li>
 <li id="l-96" _a="darobin" _r="sangwhan"><a href="#l-96">#</a> [10:43] &lt;darobin&gt; sangwhan: actually it's probably a test calling a SPARQL endpoint over the internets; but certainly not browser</li>
 <li id="l-97" _a="sangwhan"><a href="#l-97">#</a> [10:44] &lt;sangwhan&gt; so i have to be a massively intrusive person and steal the commits from his/her repo and open a new pr and suggest to close the old one i assume...</li>
 <li id="l-98" _a="jgraham" _r="Yes"><a href="#l-98">#</a> [10:44] &lt;jgraham&gt; Yes, I hate GitHub</li>
 <li id="l-99" _a="sangwhan" _r="darobin"><a href="#l-99">#</a> [10:44] &lt;sangwhan&gt; darobin: i skimmed through the spec, i think i need something to calm myself down</li>
 <li id="l-100" _a="sangwhan" _r="sadly"><a href="#l-100">#</a> [10:45] &lt;sangwhan&gt; sadly, everyone else loves GH</li>
 <li id="l-101" _a="Ms2ger"><a href="#l-101">#</a> [10:46] &lt;Ms2ger&gt; I never figured out why</li>
 <li id="l-102" _a="sangwhan" _r="jgraham"><a href="#l-102">#</a> [10:47] &lt;sangwhan&gt; jgraham: i might have asked this before, but any chance critic@hoppipolla will get support for FiddleAndTweak?</li>
 <li id="l-103" _a="jgraham" _r="sangwhan"><a href="#l-103">#</a> [10:47] &lt;jgraham&gt; sangwhan: I don't think it can</li>
 <li id="l-104" _a="jgraham"><a href="#l-104">#</a> [10:47] &lt;jgraham&gt; Evn if I enabled extensions, it wouldn't have the right permissions to push to the original repo</li>
 <li id="l-105" _a="sangwhan" _r="jgraham"><a href="#l-105">#</a> [10:48] &lt;sangwhan&gt; jgraham: that makes me a sad panda, i guess that's the downside of being dependent on github</li>
 <li id="l-106" _a="sangwhan"><a href="#l-106">#</a> [10:48] * sangwhan &lt;3 history rewrites</li>
 <li id="l-107" _a="Ms2ger"><a href="#l-107">#</a> [10:50] * Ms2ger disapproves of rewriting history</li>
 <li id="l-108" _a="sangwhan"><a href="#l-108">#</a> [10:50] * sangwhan assumes Ms2ger never writes profanity in commit messages</li>
 <li id="l-109" _a="Ms2ger"><a href="#l-109">#</a> [10:52] &lt;Ms2ger&gt; I avoid it :)</li>
 <li id="l-110" _a="jgraham" _r="Hmm"><a href="#l-110">#</a> [10:52] &lt;jgraham&gt; Hmm, there's nothing wrong with rewriting history</li>
 <li id="l-111" _a="jgraham"><a href="#l-111">#</a> [10:53] &lt;jgraham&gt; It's a very positive thing</li>
 <li id="l-112" _a="sangwhan"><a href="#l-112">#</a> [10:53] &lt;sangwhan&gt; Depends on your workflow, but if you rewrite history on master you should be shot</li>
 <li id="l-113" _a="jgraham"><a href="#l-113">#</a> [10:53] &lt;jgraham&gt; Rewriting *published* history is the only problematic thing</li>
 <li id="l-114" _a="jgraham"><a href="#l-114">#</a> [10:53] &lt;jgraham&gt; Sure</li>
 <li id="l-115" _a="Ms2ger" _r="sangwhan"><a href="#l-115">#</a> [10:53] &lt;Ms2ger&gt; sangwhan, oh :(</li>
 <li id="l-116" _a="Ms2ger"><a href="#l-116">#</a> [10:53] &lt;Ms2ger&gt; I was thinking of doing that recently</li>
 <li id="l-117" _a="sangwhan"><a href="#l-117">#</a> [10:54] &lt;sangwhan&gt; Opera style history rewriting is fine, but if you rewrite history on a shared branch with a backlog everyone has you should be shot</li>
 <li id="l-118" _a="jgraham" _r="Yeah"><a href="#l-118">#</a> [10:56] &lt;jgraham&gt; Yeah, in general the way Opera use VCS is very smooth and productive</li>
 <li id="l-119" _a="sangwhan"><a href="#l-119">#</a> [10:56] &lt;sangwhan&gt; (&quot;Opera style&quot; is doing it on a seperate repo, isolated branch, rewriting there, and merging it into the main repo's shared branch)</li>
 <li id="l-120" _a="sangwhan"><a href="#l-120">#</a> [10:56] &lt;sangwhan&gt; downside - it has a learning curve..</li>
 <li id="l-121" _a="sangwhan" _r="jgraham"><a href="#l-121">#</a> [10:57] &lt;sangwhan&gt; jgraham: darobin any test wishlists? going to make a PR for the page changes in about a hour</li>
 <li id="l-122" _a="sangwhan"><a href="#l-122">#</a> [10:57] &lt;sangwhan&gt; apart from &quot;don't write crap tests&quot;</li>
 <li id="l-123" _a="Ms2ger" _r="sangwhan"><a href="#l-123">#</a> [11:05] &lt;Ms2ger&gt; sangwhan, yeah, somewhere</li>
 <li id="l-124" _a="Ms2ger"><a href="#l-124">#</a> [11:06] &lt;Ms2ger&gt; <a href="https://etherpad.mozilla.org/VkkW1BU54C">https://etherpad.mozilla.org/VkkW1BU54C</a></li>
 <li id="l-125" _a="sangwhan" _r="Ms2ger"><a href="#l-125">#</a> [11:07] &lt;sangwhan&gt; Ms2ger: thanks, looks like this is more than enough</li>
 <li id="l-126" _a="sangwhan"><a href="#l-126">#</a> [11:07] * sangwhan blankly stares at the line &quot;* Rewrite the tests in [8] from the WebKit repository to be in testharness.js format&quot;</li>
 <li id="l-127"><a href="#l-127">#</a> [11:20] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-128" _a="abarsto"><a href="#l-128">#</a> [11:20] * abarsto is now known as ArtB</li>
 <li id="l-129"><a href="#l-129">#</a> [11:27] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-130"><a href="#l-130">#</a> [11:27] -gitbot:#testing- [web-platform-tests] AFBarstow closed pull request #794: This is the test suite for the Touch Events spec &lt;<a href="http://www.w3.org/TR/t">http://www.w3.org/TR/t</a>... (master...AFBarstow/touch-events) <a href="https://github.com/w3c/web-platform-tests/pull/794">https://github.com/w3c/web-platform-tests/pull/794</a></li>
 <li id="l-131"><a href="#l-131">#</a> [11:27] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-132"><a href="#l-132">#</a> [11:42] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-133"><a href="#l-133">#</a> [12:00] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-134"><a href="#l-134">#</a> [12:00] * Joins: MK (kawamori@public.cloak)</li>
 <li id="l-135"><a href="#l-135">#</a> [12:08] * Quits: MK (kawamori@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-136" _a="jgraham" _r="sangwhan"><a href="#l-136">#</a> [12:11] &lt;jgraham&gt; sangwhan: That would actually be one of the more useful things in that document</li>
 <li id="l-137" _a="sangwhan" _r="jgraham"><a href="#l-137">#</a> [12:20] &lt;sangwhan&gt; jgraham: it's also not a trivial task</li>
 <li id="l-138" _a="jgraham"><a href="#l-138">#</a> [12:20] &lt;jgraham&gt; No</li>
 <li id="l-139" _a="jgraham"><a href="#l-139">#</a> [12:21] &lt;jgraham&gt; But writing good tests generally isn't</li>
 <li id="l-140" _a="Ms2ger"><a href="#l-140">#</a> [12:22] &lt;Ms2ger&gt; Who'd have thought</li>
 <li id="l-141" _a="jgraham"><a href="#l-141">#</a> [12:22] &lt;jgraham&gt; Unfortunately this exposes the main failing of TTWF</li>
 <li id="l-142" _a="jgraham"><a href="#l-142">#</a> [12:23] &lt;jgraham&gt; So far we have had roughly a 0% retention rate</li>
 <li id="l-143" _a="jgraham"><a href="#l-143">#</a> [12:23] &lt;jgraham&gt; And although we have got some good tests it hasn't been that many compared to the number of first-time participants</li>
 <li id="l-144" _a="sangwhan"><a href="#l-144">#</a> [12:24] &lt;sangwhan&gt; low retention rate doesn't surprise me at all, writing tests is not fun after all</li>
 <li id="l-145" _a="jgraham" _r="Umm"><a href="#l-145">#</a> [12:24] &lt;jgraham&gt; Umm, it can be</li>
 <li id="l-146" _a="jgraham"><a href="#l-146">#</a> [12:25] &lt;jgraham&gt; I think it depends on the situation</li>
 <li id="l-147" _a="sangwhan"><a href="#l-147">#</a> [12:25] &lt;sangwhan&gt; one way to try to increase retention rate might be to try to create some sort of local discussion/interest group and herd the people who are interested there, want me to try that?</li>
 <li id="l-148" _a="jgraham"><a href="#l-148">#</a> [12:25] &lt;jgraham&gt; I don't really enjoy writing tests for my own code</li>
 <li id="l-149" _a="sangwhan"><a href="#l-149">#</a> [12:25] &lt;sangwhan&gt; my code never needs tests, it's always perfect (cough)</li>
 <li id="l-150" _a="jgraham"><a href="#l-150">#</a> [12:25] &lt;jgraham&gt; But writing tests with the explicit aim of finding bugs in someone else's code is quite different</li>
 <li id="l-151" _a="jgraham"><a href="#l-151">#</a> [12:25] &lt;jgraham&gt; That could be interesting</li>
 <li id="l-152" _a="sangwhan"><a href="#l-152">#</a> [12:26] &lt;sangwhan&gt; lemme put that on the list of things to bring up during the event then</li>
 <li id="l-153" _a="jgraham"><a href="#l-153">#</a> [12:27] &lt;jgraham&gt; I think one problem we've had in the past is that we've concentrated hard on number of participants rather than on finding participants who are unusually likely to have the right skillset</li>
 <li id="l-154" _a="jgraham"><a href="#l-154">#</a> [12:27] &lt;jgraham&gt; I think it would be a neat idea to host a TTWF at a major js conf. or similar</li>
 <li id="l-155" _a="sangwhan" _r="yes"><a href="#l-155">#</a> [12:29] &lt;sangwhan&gt; yes, i agree on that bit - tokyo was actually pretty good imho, but seoul is a bit risky as it's a extension of some other event that focuses more on academia related matters (search for www2014)</li>
 <li id="l-156" _a="sangwhan" _r="also"><a href="#l-156">#</a> [12:29] &lt;sangwhan&gt; also: <a href="https://etherpad.mozilla.org/sHzr5W5imP">https://etherpad.mozilla.org/sHzr5W5imP</a></li>
 <li id="l-157" _a="jgraham"><a href="#l-157">#</a> [12:32] &lt;jgraham&gt; I think getting people to review is a bit hopeless</li>
 <li id="l-158" _a="Ms2ger"><a href="#l-158">#</a> [12:33] &lt;Ms2ger&gt; Right</li>
 <li id="l-159" _a="Ms2ger" _r="sangwhan"><a href="#l-159">#</a> [12:34] &lt;Ms2ger&gt; sangwhan, if you're there, please do ask people to reply to review comments, even if they come in after the event is done</li>
 <li id="l-160" _a="Ms2ger"><a href="#l-160">#</a> [12:34] &lt;Ms2ger&gt; And make them add their email address on critic</li>
 <li id="l-161" _a="sangwhan" _r="Ms2ger"><a href="#l-161">#</a> [12:35] &lt;sangwhan&gt; Ms2ger: i have always asked for the people i have helped in place, but what for some reason &quot;pretty please with sugar on top&quot; wasn't enough for that to happen</li>
 <li id="l-162" _a="sangwhan" _r="jgraham"><a href="#l-162">#</a> [12:35] &lt;sangwhan&gt; jgraham: ok, i'll scratch that</li>
 <li id="l-163" _a="Ms2ger"><a href="#l-163">#</a> [12:36] &lt;Ms2ger&gt; Maybe during the initial presentation, stop and tell them you'll continue when they've done it :)</li>
 <li id="l-164" _a="sangwhan"><a href="#l-164">#</a> [12:36] &lt;sangwhan&gt; I'll have to note that most people hate critic</li>
 <li id="l-165" _a="jgraham"><a href="#l-165">#</a> [12:37] &lt;jgraham&gt; That isn't my experience</li>
 <li id="l-166" _a="sangwhan"><a href="#l-166">#</a> [12:37] &lt;sangwhan&gt; It's targeted for a very narrow spectrum of audience, and it's a pretty hard sell if you come from a background where Github is the best thing after Google Code SVN</li>
 <li id="l-167" _a="MikeSmith"><a href="#l-167">#</a> [12:37] &lt;MikeSmith&gt; are there no appcache tests in presto-testo?</li>
 <li id="l-168" _a="jgraham"><a href="#l-168">#</a> [12:38] &lt;jgraham&gt; Most people think that it's ugly and confusing for a while and then appreciate it</li>
 <li id="l-169" _a="sangwhan" _r="jgraham"><a href="#l-169">#</a> [12:38] &lt;sangwhan&gt; jgraham: maybe you and i i think have been dealing with a similar demographic?</li>
 <li id="l-170" _a="jgraham"><a href="#l-170">#</a> [12:38] &lt;jgraham&gt; The problem is people that never get past the first stage</li>
 <li id="l-171" _a="sangwhan" _r="MikeSmith"><a href="#l-171">#</a> [12:38] &lt;sangwhan&gt; MikeSmith: good question, find -name doesn't give you anything?</li>
 <li id="l-172" _a="MikeSmith"><a href="#l-172">#</a> [12:38] &lt;MikeSmith&gt; nope</li>
 <li id="l-173" _a="MikeSmith"><a href="#l-173">#</a> [12:39] &lt;MikeSmith&gt; for &quot;*cache*&quot; or &quot;*offline*&quot;</li>
 <li id="l-174" _a="MikeSmith" _r="sangwhan"><a href="#l-174">#</a> [12:39] &lt;MikeSmith&gt; sangwhan: do you know like critic yourself?</li>
 <li id="l-175"><a href="#l-175">#</a> [12:40] * Quits: Ms2ger (~Ms2ger@public.cloak) (&quot;bbl&quot;)</li>
 <li id="l-176" _a="MikeSmith"><a href="#l-176">#</a> [12:40] &lt;MikeSmith&gt; I wonder what other better real code-review systems exist that are any better</li>
 <li id="l-177" _a="MikeSmith"><a href="#l-177">#</a> [12:40] &lt;MikeSmith&gt; I know many that are lot worse</li>
 <li id="l-178" _a="jgraham" _r="Right"><a href="#l-178">#</a> [12:40] &lt;jgraham&gt; Right, my experience is that critic is the worst code review system apart from all the others</li>
 <li id="l-179" _a="MikeSmith"><a href="#l-179">#</a> [12:41] &lt;MikeSmith&gt; yeah</li>
 <li id="l-180" _a="MikeSmith"><a href="#l-180">#</a> [12:41] &lt;MikeSmith&gt; I think somebody who's never had to use a real code-review system are never going to like the first one they use the first times they use it</li>
 <li id="l-181" _a="MikeSmith"><a href="#l-181">#</a> [12:42] &lt;MikeSmith&gt; but then after they use that one for 3 months they will claim it's the best one there is and every one they try after that they'll hate more</li>
 <li id="l-182"><a href="#l-182">#</a> [12:43] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-183"><a href="#l-183">#</a> [12:44] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-184" _a="sangwhan" _r="MikeSmith"><a href="#l-184">#</a> [12:45] &lt;sangwhan&gt; MikeSmith: while i have found some tests internally, seems like those tests are not a part of what is called presto-testo</li>
 <li id="l-185" _a="MikeSmith"><a href="#l-185">#</a> [12:45] &lt;MikeSmith&gt; ok</li>
 <li id="l-186" _a="MikeSmith"><a href="#l-186">#</a> [12:46] &lt;MikeSmith&gt; well we have some tests that infraware submitted</li>
 <li id="l-187" _a="sangwhan" _r="yeah"><a href="#l-187">#</a> [12:46] &lt;sangwhan&gt; yeah, that could do</li>
 <li id="l-188" _a="MikeSmith"><a href="#l-188">#</a> [12:46] &lt;MikeSmith&gt; which are not total disasters</li>
 <li id="l-189" _a="MikeSmith"><a href="#l-189">#</a> [12:46] &lt;MikeSmith&gt; but nobody has really reviewed them yet</li>
 <li id="l-190" _a="sangwhan"><a href="#l-190">#</a> [12:46] &lt;sangwhan&gt; odinho might know better</li>
 <li id="l-191" _a="MikeSmith"><a href="#l-191">#</a> [12:46] &lt;MikeSmith&gt; ok</li>
 <li id="l-192" _a="jgraham"><a href="#l-192">#</a> [12:47] &lt;jgraham&gt; I think appcache has the problem that it's hard to write automatic tests for</li>
 <li id="l-193" _a="MikeSmith"><a href="#l-193">#</a> [12:47] &lt;MikeSmith&gt; denis has reviewed the infraware ones and sanified them some and opened a PR but nobody's reviewed that PR yet</li>
 <li id="l-194" _a="jgraham"><a href="#l-194">#</a> [12:47] &lt;jgraham&gt; Because you need a way to go offline</li>
 <li id="l-195" _a="MikeSmith"><a href="#l-195">#</a> [12:47] &lt;MikeSmith&gt; yeah</li>
 <li id="l-196" _a="jgraham"><a href="#l-196">#</a> [12:48] &lt;jgraham&gt; I guess WebDriver doesn't have a way to do that, but maybe it should have</li>
 <li id="l-197" _a="jgraham"><a href="#l-197">#</a> [12:48] &lt;jgraham&gt; It seems like it would be super useful for our testing and for testing actual offline apps</li>
 <li id="l-198" _a="jgraham" _r="AutomatedTester"><a href="#l-198">#</a> [12:48] &lt;jgraham&gt; AutomatedTester: ^</li>
 <li id="l-199" _a="jgraham"><a href="#l-199">#</a> [12:50] &lt;jgraham&gt; In other news, testharness.js has a problem</li>
 <li id="l-200"><a href="#l-200">#</a> [12:51] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-201" _a="jgraham"><a href="#l-201">#</a> [12:51] &lt;jgraham&gt; If a test causes an exception that isn't caught it sets the harness status to Error and aborts the run</li>
 <li id="l-202" _a="jgraham"><a href="#l-202">#</a> [12:51] &lt;jgraham&gt; But aborting the run tends to lead to race conditions</li>
 <li id="l-203" _a="jgraham"><a href="#l-203">#</a> [12:52] &lt;jgraham&gt; e.g. if you have two async tests and one raises a bogus exception then the other one may or may not complete</li>
 <li id="l-204" _a="jgraham"><a href="#l-204">#</a> [12:52] &lt;jgraham&gt; So I wonder if it should set the status to error but let the other tests keep running as long as possible</li>
 <li id="l-205" _a="jgraham"><a href="#l-205">#</a> [12:53] &lt;jgraham&gt; It isn't possible to be *sure* that won't lead to races, but it at least makes it a bit less likely</li>
 <li id="l-206" _a="MikeSmith"><a href="#l-206">#</a> [12:53] &lt;MikeSmith&gt; speaking of appcache, Somebody should respond to <a href="http://lists.whatwg.org/htdig.cgi/whatwg-whatwg.org/2014-March/084588.html">http://lists.whatwg.org/htdig.cgi/whatwg-whatwg.org/2014-March/084588.html</a></li>
 <li id="l-207" _a="jgraham" _r="zcorpan"><a href="#l-207">#</a> [12:53] &lt;jgraham&gt; zcorpan: ^</li>
 <li id="l-208" _a="jgraham"><a href="#l-208">#</a> [12:53] &lt;jgraham&gt; (the tests thing, not the appecache thing)</li>
 <li id="l-209" _a="zcorpan" _r="jgraham"><a href="#l-209">#</a> [12:56] &lt;zcorpan&gt; jgraham: bogus exception here is one that isn't caught by step() ?</li>
 <li id="l-210" _a="jgraham"><a href="#l-210">#</a> [12:57] &lt;jgraham&gt; Yes</li>
 <li id="l-211" _a="jgraham"><a href="#l-211">#</a> [12:57] &lt;jgraham&gt; So in this case a worker is raising an error that is propogating to Window</li>
 <li id="l-212" _a="zcorpan" _r="jgraham"><a href="#l-212">#</a> [12:57] &lt;zcorpan&gt; jgraham: how do you know which test is responsible?</li>
 <li id="l-213" _a="jgraham"><a href="#l-213">#</a> [12:57] &lt;jgraham&gt; But the step() can't catch that because we don't have step() inside the worker</li>
 <li id="l-214" _a="jgraham" _r="zcorpan"><a href="#l-214">#</a> [12:58] &lt;jgraham&gt; zcorpan: You don't</li>
 <li id="l-215" _a="jgraham"><a href="#l-215">#</a> [12:58] &lt;jgraham&gt; (in general, in this case I do)</li>
 <li id="l-216" _a="zcorpan"><a href="#l-216">#</a> [12:58] &lt;zcorpan&gt; yeah i meant how does testharness know :-)</li>
 <li id="l-217" _a="zcorpan"><a href="#l-217">#</a> [12:58] &lt;zcorpan&gt; ok</li>
 <li id="l-218" _a="jgraham"><a href="#l-218">#</a> [12:59] &lt;jgraham&gt; *Hopefully* if the test is raising an exception it doesn't come back as PASS</li>
 <li id="l-219" _a="zcorpan"><a href="#l-219">#</a> [12:59] &lt;zcorpan&gt; so what's the idea, set global result to &quot;error&quot; but continue running in case some tests carry on fine?</li>
 <li id="l-220" _a="jgraham"><a href="#l-220">#</a> [12:59] &lt;jgraham&gt; Yeah</li>
 <li id="l-221" _a="jgraham"><a href="#l-221">#</a> [12:59] &lt;jgraham&gt; Well specifically in case some tests have already started</li>
 <li id="l-222"><a href="#l-222">#</a> [12:59] * Joins: plh (plehegar@public.cloak)</li>
 <li id="l-223" _a="zcorpan"><a href="#l-223">#</a> [12:59] &lt;zcorpan&gt; that seems like a good idea. i guess it increases the run time but this seems like a rare thing anyway</li>
 <li id="l-224" _a="jgraham"><a href="#l-224">#</a> [12:59] &lt;jgraham&gt; In which case abruptly stopping them is racy</li>
 <li id="l-225" _a="jgraham" _r="zcorpan"><a href="#l-225">#</a> [13:02] &lt;jgraham&gt; zcorpan: Totally unrelated, but when you are in Linköping next there should be chocolate for you</li>
 <li id="l-226" _a="zcorpan" _r="jgraham"><a href="#l-226">#</a> [13:02] &lt;zcorpan&gt; jgraham: oooooh!</li>
 <li id="l-227" _a="zcorpan"><a href="#l-227">#</a> [13:02] &lt;zcorpan&gt; on monday actually</li>
 <li id="l-228" _a="zcorpan"><a href="#l-228">#</a> [13:02] &lt;zcorpan&gt; thanks!</li>
 <li id="l-229" _a="jgraham" _r="Hmm"><a href="#l-229">#</a> [13:06] &lt;jgraham&gt; Hmm, so that solved one problem</li>
 <li id="l-230" _a="jgraham"><a href="#l-230">#</a> [13:07] &lt;jgraham&gt; But it's still racy whether or not I get the harness ERROR before all the other tests are finished</li>
 <li id="l-231" _a="jgraham"><a href="#l-231">#</a> [13:09] &lt;jgraham&gt; I don't have any good ideas on how to solve that</li>
 <li id="l-232" _a="odinho" _r="No"><a href="#l-232">#</a> [14:22] &lt;odinho&gt; No, I didn't look very much into AppCache, so I don't know.</li>
 <li id="l-233" _a="odinho"><a href="#l-233">#</a> [14:23] &lt;odinho&gt; Saw that only very few of my tests were in jgraham's flaky list though.  They were more flaky before, but running it in Spartan (Opera Test system that jgraham worked on) nagged me about their unstableness so they were stable in Spartan in the end.</li>
 <li id="l-234" _a="odinho"><a href="#l-234">#</a> [14:24] &lt;odinho&gt; So I guess that's why only a very few popped up here now.  --  Not immediately obvious why though (haven't looked at anything else than the report)</li>
 <li id="l-235" _a="odinho"><a href="#l-235">#</a> [14:27] &lt;odinho&gt; Microsoft-idb-port is 100% now. With 37 issues. Work to be done.</li>
 <li id="l-236" _a="jgraham" _r="odinho"><a href="#l-236">#</a> [14:33] &lt;jgraham&gt; odinho: Nice</li>
 <li id="l-237" _a="odinho" _r="jgraham"><a href="#l-237">#</a> [14:34] &lt;odinho&gt; jgraham: all zqzhang, lest you think it was me. :P</li>
 <li id="l-238" _a="jgraham"><a href="#l-238">#</a> [14:34] &lt;jgraham&gt; And he's not even here so I can thank him :(</li>
 <li id="l-239"><a href="#l-239">#</a> [14:38] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-240"><a href="#l-240">#</a> [14:56] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-241"><a href="#l-241">#</a> [14:57] * Joins: anssik (~uid10742@public.cloak)</li>
 <li id="l-242"><a href="#l-242">#</a> [14:59] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-243"><a href="#l-243">#</a> [14:59] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-244"><a href="#l-244">#</a> [14:59] -gitbot:#testing- [web-platform-tests] zcorpan closed pull request #712: Tests for the named access on the window object (master...submission/xiaojunwu/window-named-objects) <a href="https://github.com/w3c/web-platform-tests/pull/712">https://github.com/w3c/web-platform-tests/pull/712</a></li>
 <li id="l-245"><a href="#l-245">#</a> [14:59] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-246"><a href="#l-246">#</a> [15:16] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-247"><a href="#l-247">#</a> [15:41] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-248"><a href="#l-248">#</a> [15:59] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-249"><a href="#l-249">#</a> [15:59] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-250"><a href="#l-250">#</a> [16:03] * Quits: AutomatedTester (~AutomatedTester@public.cloak) (Client closed connection)</li>
 <li id="l-251"><a href="#l-251">#</a> [16:03] * Joins: AutomatedTester (~AutomatedTester@public.cloak)</li>
 <li id="l-252"><a href="#l-252">#</a> [16:10] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-253"><a href="#l-253">#</a> [16:10] -gitbot:#testing- [web-platform-tests] dontcallmedom closed pull request #376: Submission/ldeluca (master...submission/ldeluca) <a href="https://github.com/w3c/web-platform-tests/pull/376">https://github.com/w3c/web-platform-tests/pull/376</a></li>
 <li id="l-254"><a href="#l-254">#</a> [16:10] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-255"><a href="#l-255">#</a> [16:15] * Quits: scott_gonzalez (~scott_gonzalez@public.cloak) (scott_gonzalez)</li>
 <li id="l-256"><a href="#l-256">#</a> [16:19] * Joins: lukeis (~Adium@public.cloak)</li>
 <li id="l-257"><a href="#l-257">#</a> [16:24] * Joins: kawai (~kawai@public.cloak)</li>
 <li id="l-258"><a href="#l-258">#</a> [16:24] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-259"><a href="#l-259">#</a> [16:26] * Quits: ArtB (~abarsto@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-260"><a href="#l-260">#</a> [16:30] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-261"><a href="#l-261">#</a> [16:30] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/worker_terminate_fixup (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/eb67b371f0b9">https://github.com/w3c/web-platform-tests/commit/eb67b371f0b9</a></li>
 <li id="l-262"><a href="#l-262">#</a> [16:30] -gitbot:#testing- web-platform-tests/jgraham/worker_terminate_fixup eb67b37 James Graham: Change terminate test not to spam the event loop with so many messages</li>
 <li id="l-263"><a href="#l-263">#</a> [16:30] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-264"><a href="#l-264">#</a> [16:31] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-265"><a href="#l-265">#</a> [16:31] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #804: Change terminate test not to spam the event loop with so many messages (master...jgraham/worker_terminate_fixup) <a href="https://github.com/w3c/web-platform-tests/pull/804">https://github.com/w3c/web-platform-tests/pull/804</a></li>
 <li id="l-266"><a href="#l-266">#</a> [16:31] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-267"><a href="#l-267">#</a> [16:49] * Joins: lmclister (~lmclister@public.cloak)</li>
 <li id="l-268"><a href="#l-268">#</a> [16:55] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-269"><a href="#l-269">#</a> [16:55] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-270"><a href="#l-270">#</a> [17:02] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-271"><a href="#l-271">#</a> [17:20] * Quits: ptressel (~chatzilla@public.cloak) (&quot;zzz&quot;)</li>
 <li id="l-272"><a href="#l-272">#</a> [17:26] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-273"><a href="#l-273">#</a> [17:28] * Joins: zcorpan_ (~zcorpan@public.cloak)</li>
 <li id="l-274"><a href="#l-274">#</a> [17:28] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-275"><a href="#l-275">#</a> [17:30] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
 <li id="l-276"><a href="#l-276">#</a> [17:35] * Quits: zcorpan_ (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-277" _a="odinho"><a href="#l-277">#</a> [17:36] &lt;odinho&gt; jgraham + everyone: So in the IDB tests. If indexedDB doesn't exist. Should I just /not/ do something. Or assert_unreached(&quot;tis' not supported&quot;)?</li>
 <li id="l-278" _a="jgraham" _r="odinho"><a href="#l-278">#</a> [17:37] &lt;jgraham&gt; odinho: Well</li>
 <li id="l-279" _a="jgraham"><a href="#l-279">#</a> [17:37] &lt;jgraham&gt; It's good to still run all the tests</li>
 <li id="l-280" _a="jgraham"><a href="#l-280">#</a> [17:37] &lt;jgraham&gt; Or at least still declare all the tests</li>
 <li id="l-281" _a="jgraham"><a href="#l-281">#</a> [17:37] &lt;jgraham&gt; But it's OK to bail early within a test</li>
 <li id="l-282" _a="jgraham"><a href="#l-282">#</a> [17:38] &lt;jgraham&gt; so doing test(function() {assert_true(IndexedDB in window); test(function(){ /*only run this test if the other one passed})}) is not really OK</li>
 <li id="l-283" _a="jgraham"><a href="#l-283">#</a> [17:40] &lt;jgraham&gt; But doing test(function() {var supported = IndexedDB in window; assert_true(supported)}); test(function() {assert_true(supported); /*rest of test*/}) is fine</li>
 <li id="l-284"><a href="#l-284">#</a> [17:44] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-285"><a href="#l-285">#</a> [17:44] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/websocket_wss_disable (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/4c7514843600">https://github.com/w3c/web-platform-tests/commit/4c7514843600</a></li>
 <li id="l-286"><a href="#l-286">#</a> [17:44] -gitbot:#testing- web-platform-tests/jgraham/websocket_wss_disable 4c75148 James Graham: Don't pretend like we support wss for now</li>
 <li id="l-287"><a href="#l-287">#</a> [17:44] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-288"><a href="#l-288">#</a> [17:45] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-289"><a href="#l-289">#</a> [17:45] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #805: Don't pretend like we support wss for now (master...jgraham/websocket_wss_disable) <a href="https://github.com/w3c/web-platform-tests/pull/805">https://github.com/w3c/web-platform-tests/pull/805</a></li>
 <li id="l-290"><a href="#l-290">#</a> [17:45] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-291" _a="glenn"><a href="#l-291">#</a> [17:45] &lt;glenn&gt; random input: it would be nice feature on test runner to deselect tests known to regularly time out on the platform under test, e.g., by using a previously generated  (and updated) JSON file containing a list of such tests</li>
 <li id="l-292" _a="odinho" _r="jgraham"><a href="#l-292">#</a> [17:46] &lt;odinho&gt; jgraham: Yeah, -- this was not really a question about that. It was if I should let the result be NOT RUN or FAILED.</li>
 <li id="l-293" _a="jgraham" _r="glenn"><a href="#l-293">#</a> [17:47] &lt;jgraham&gt; glenn: Yeah having a way to select exactly which tests to include/exclude could be be useful. Patches welcome, I guess.</li>
 <li id="l-294" _a="jgraham" _r="odinho"><a href="#l-294">#</a> [17:47] &lt;jgraham&gt; odinho: I don't think there's a preference</li>
 <li id="l-295" _a="odinho"><a href="#l-295">#</a> [17:47] &lt;odinho&gt; I'm creating async_test anyway, I could just never actually start it to get not run, -- though I don't really like it to always time out. I can just fail it directly.</li>
 <li id="l-296" _a="odinho" _r="OK"><a href="#l-296">#</a> [17:47] &lt;odinho&gt; OK, so I'll fail.</li>
 <li id="l-297" _a="odinho" _r="And"><a href="#l-297">#</a> [17:48] &lt;odinho&gt; And, it's been long since I did testing, -- what's the ideomatic JS way to test whether something is truthish? I have it as a precursor to quite a few tests, because I don't want them to pass if the exception they get is due to some other reason.</li>
 <li id="l-298" _a="glenn" _r="jgraham"><a href="#l-298">#</a> [17:48] &lt;glenn&gt; jgraham: i'm beginning a project that will have me using the WPT regularly, so perhaps I'll create and post a pull request; it would certainly help for smoke test scenarios; when i ran with the WPT runner recently it spent a couple of hours (seemingly) in time out mode</li>
 <li id="l-299" _a="odinho"><a href="#l-299">#</a> [17:48] &lt;odinho&gt; assert_true(!!cursor)  &lt;?</li>
 <li id="l-300" _a="jgraham" _r="odinho"><a href="#l-300">#</a> [17:49] &lt;jgraham&gt; odinho: Yeah, I suppose if you don't already know that it's actually a bool</li>
 <li id="l-301" _a="jgraham"><a href="#l-301">#</a> [17:49] &lt;jgraham&gt; But I might question why you don't know</li>
 <li id="l-302" _a="odinho"><a href="#l-302">#</a> [17:51] &lt;odinho&gt; It's not a bool.</li>
 <li id="l-303" _a="odinho"><a href="#l-303">#</a> [17:51] &lt;odinho&gt; It's a IDBCursor or null.</li>
 <li id="l-304" _a="odinho"><a href="#l-304">#</a> [17:51] &lt;odinho&gt; Guess that's a good answer then :P</li>
 <li id="l-305" _a="odinho"><a href="#l-305">#</a> [17:53] &lt;odinho&gt; assert_true(cursor instanceof IDBCursor)  nicer</li>
 <li id="l-306"><a href="#l-306">#</a> [17:53] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-307" _a="jgraham"><a href="#l-307">#</a> [17:53] &lt;jgraham&gt; Yeah or !== null if you might have cross-window IDBCursor</li>
 <li id="l-308"><a href="#l-308">#</a> [17:54] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-309"><a href="#l-309">#</a> [18:00] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
 <li id="l-310" _a="jgraham" _r="Ms2ger"><a href="#l-310">#</a> [18:01] &lt;jgraham&gt; Ms2ger: !</li>
 <li id="l-311"><a href="#l-311">#</a> [18:01] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-312" _a="jgraham" _r="Ms2ger"><a href="#l-312">#</a> [18:01] &lt;jgraham&gt; Ms2ger: <a href="https://critic.hoppipolla.co.uk/showcomment?chain=3515">https://critic.hoppipolla.co.uk/showcomment?chain=3515</a></li>
 <li id="l-313"><a href="#l-313">#</a> [18:02] * Quits: abarsto (~abarsto@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-314" _a="Ms2ger"><a href="#l-314">#</a> [18:07] &lt;Ms2ger&gt; I sure hope that isn't innerHTML again :)</li>
 <li id="l-315" _a="Ms2ger"><a href="#l-315">#</a> [18:07] * Ms2ger kicks his mouse</li>
 <li id="l-316" _a="Ms2ger"><a href="#l-316">#</a> [18:07] &lt;Ms2ger&gt; brb</li>
 <li id="l-317" _a="Ms2ger"><a href="#l-317">#</a> [18:08] &lt;Ms2ger&gt; back</li>
 <li id="l-318" _a="jgraham"><a href="#l-318">#</a> [18:08] &lt;jgraham&gt; Not innerHTML</li>
 <li id="l-319" _a="jgraham"><a href="#l-319">#</a> [18:08] &lt;jgraham&gt; Although now you mention it…</li>
 <li id="l-320" _a="Ms2ger"><a href="#l-320">#</a> [18:09] &lt;Ms2ger&gt; I know, right</li>
 <li id="l-321" _a="Ms2ger"><a href="#l-321">#</a> [18:09] &lt;Ms2ger&gt; Is that the workers intermittent timeout?</li>
 <li id="l-322" _a="jgraham" _r="Ms2ger"><a href="#l-322">#</a> [18:09] &lt;jgraham&gt; Ms2ger: Yeah, but specifically I wonder if you have any ideas for the issue in that comment</li>
 <li id="l-323" _a="Ms2ger"><a href="#l-323">#</a> [18:11] &lt;Ms2ger&gt; s/before the test fails/before the test finishes/?</li>
 <li id="l-324" _a="Ms2ger"><a href="#l-324">#</a> [18:12] &lt;Ms2ger&gt; I don't have any bright ideas right now</li>
 <li id="l-325" _a="jgraham"><a href="#l-325">#</a> [18:13] &lt;jgraham&gt; Yeah s/fails/finishes/ if you like</li>
 <li id="l-326"><a href="#l-326">#</a> [18:17] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-327"><a href="#l-327">#</a> [18:17] -gitbot:#testing- [web-platform-tests] Velmont pushed 1 new commit to Velmont/move_ported_idb_tests: <a href="https://github.com/w3c/web-platform-tests/commit/024366e5fb2104cee4490e25befa6f4bab28f714">https://github.com/w3c/web-platform-tests/commit/024366e5fb2104cee4490e25befa6f4bab28f714</a></li>
 <li id="l-328"><a href="#l-328">#</a> [18:17] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 024366e Odin Hørthe Omdal: IndexedDB review fixes</li>
 <li id="l-329"><a href="#l-329">#</a> [18:17] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-330"><a href="#l-330">#</a> [18:23] * Joins: abarsto (~abarsto@public.cloak)</li>
 <li id="l-331" _a="abarsto"><a href="#l-331">#</a> [18:23] * abarsto is now known as ArtB</li>
 <li id="l-332"><a href="#l-332">#</a> [18:23] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
 <li id="l-333"><a href="#l-333">#</a> [18:24] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-334"><a href="#l-334">#</a> [18:27] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-335"><a href="#l-335">#</a> [18:29] * Quits: kawai (~kawai@public.cloak) (Client closed connection)</li>
 <li id="l-336"><a href="#l-336">#</a> [18:30] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-337"><a href="#l-337">#</a> [18:30] -gitbot:#testing- [web-platform-tests] jgraham created jgraham/remove_external_url (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/0fa8f70f31f1">https://github.com/w3c/web-platform-tests/commit/0fa8f70f31f1</a></li>
 <li id="l-338"><a href="#l-338">#</a> [18:30] -gitbot:#testing- web-platform-tests/jgraham/remove_external_url 0fa8f70 James Graham: Using external URLs isn't allowed</li>
 <li id="l-339"><a href="#l-339">#</a> [18:30] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-340"><a href="#l-340">#</a> [18:31] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-341" _a="jgraham" _r="Ms2ger"><a href="#l-341">#</a> [18:32] &lt;jgraham&gt; Ms2ger: Maybe we'll just have to put allow_uncaucht_exceptions in a bunch of workers tests that *could* fail. I'm not sure what great alternatives there are</li>
 <li id="l-342" _a="Ms2ger"><a href="#l-342">#</a> [18:32] &lt;Ms2ger&gt; Yeah</li>
 <li id="l-343" _a="jgraham"><a href="#l-343">#</a> [18:33] * jgraham wonders what kind of punishment Lucifer has in store for people that try to depend on external servers in tests</li>
 <li id="l-344"><a href="#l-344">#</a> [18:57] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-345"><a href="#l-345">#</a> [19:01] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-346"><a href="#l-346">#</a> [19:02] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-347"><a href="#l-347">#</a> [19:02] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-348"><a href="#l-348">#</a> [19:02] * Joins: glenn_ (~gadams@public.cloak)</li>
 <li id="l-349"><a href="#l-349">#</a> [19:02] * Quits: glenn_ (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-350"><a href="#l-350">#</a> [19:03] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-351"><a href="#l-351">#</a> [19:10] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-352"><a href="#l-352">#</a> [19:29] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-353"><a href="#l-353">#</a> [19:36] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-354"><a href="#l-354">#</a> [19:55] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-355"><a href="#l-355">#</a> [19:55] -gitbot:#testing- [web-platform-tests] Velmont pushed 1 new commit to Velmont/move_ported_idb_tests: <a href="https://github.com/w3c/web-platform-tests/commit/244fd997d2b0db1279175b84f29cd3b30f5921d6">https://github.com/w3c/web-platform-tests/commit/244fd997d2b0db1279175b84f29cd3b30f5921d6</a></li>
 <li id="l-356"><a href="#l-356">#</a> [19:55] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 244fd99 Odin Hørthe Omdal: fixup! IndexedDB review fixes</li>
 <li id="l-357"><a href="#l-357">#</a> [19:55] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-358"><a href="#l-358">#</a> [20:03] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-359"><a href="#l-359">#</a> [20:05] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-360"><a href="#l-360">#</a> [20:05] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-361"><a href="#l-361">#</a> [20:06] * Quits: anssik (~uid10742@public.cloak) (&quot;Connection closed for inactivity&quot;)</li>
 <li id="l-362"><a href="#l-362">#</a> [20:24] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-363"><a href="#l-363">#</a> [20:24] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-364"><a href="#l-364">#</a> [20:25] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-365"><a href="#l-365">#</a> [20:30] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-366"><a href="#l-366">#</a> [20:31] * Joins: zcorpan_ (~zcorpan@public.cloak)</li>
 <li id="l-367"><a href="#l-367">#</a> [20:31] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-368"><a href="#l-368">#</a> [20:39] * Quits: zcorpan_ (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-369"><a href="#l-369">#</a> [20:40] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-370"><a href="#l-370">#</a> [20:40] * Quits: glenn (~gadams@public.cloak) (Client closed connection)</li>
 <li id="l-371"><a href="#l-371">#</a> [20:40] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-372"><a href="#l-372">#</a> [20:42] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-373"><a href="#l-373">#</a> [20:42] -gitbot:#testing- [web-platform-tests] Velmont pushed 2 new commits to Velmont/move_ported_idb_tests: <a href="https://github.com/w3c/web-platform-tests/compare/244fd997d2b0...6b963cf489c3">https://github.com/w3c/web-platform-tests/compare/244fd997d2b0...6b963cf489c3</a></li>
 <li id="l-374"><a href="#l-374">#</a> [20:42] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 89512d6 Odin Hørthe Omdal: fixup! IndexedDB review fixes</li>
 <li id="l-375"><a href="#l-375">#</a> [20:42] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 6b963cf Odin Hørthe Omdal: IndexedDB: Whitespace cleanup</li>
 <li id="l-376"><a href="#l-376">#</a> [20:42] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-377"><a href="#l-377">#</a> [20:47] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-378"><a href="#l-378">#</a> [20:56] * Quits: rhauck (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-379"><a href="#l-379">#</a> [20:56] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-380" _a="odinho"><a href="#l-380">#</a> [20:58] &lt;odinho&gt; omfg</li>
 <li id="l-381"><a href="#l-381">#</a> [20:58] * Quits: rhauck (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-382" _a="odinho"><a href="#l-382">#</a> [20:58] &lt;odinho&gt; That newline fix reeeeeeeeeaally messed things up</li>
 <li id="l-383" _a="odinho"><a href="#l-383">#</a> [20:58] &lt;odinho&gt; This rebase is extreme. Need to trick it I guess.</li>
 <li id="l-384"><a href="#l-384">#</a> [21:03] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-385"><a href="#l-385">#</a> [21:06] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-386"><a href="#l-386">#</a> [21:06] -gitbot:#testing- [web-platform-tests] Velmont force-pushed Velmont/move_ported_idb_tests from 6b963cf to b7261d2: <a href="https://github.com/w3c/web-platform-tests/commits/Velmont/move_ported_idb_tests">https://github.com/w3c/web-platform-tests/commits/Velmont/move_ported_idb_tests</a></li>
 <li id="l-387"><a href="#l-387">#</a> [21:06] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 200ae6f Odin Hørthe Omdal: IndexedDB: Delete the Opera-ported Microsoft IDB tests</li>
 <li id="l-388"><a href="#l-388">#</a> [21:06] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests c8092c8 Odin Hørthe Omdal: IndexedDB: Ressurect the rewritten tests</li>
 <li id="l-389"><a href="#l-389">#</a> [21:06] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 641a57b Odin Hørthe Omdal: IndexedDB review fixes</li>
 <li id="l-390"><a href="#l-390">#</a> [21:06] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-391" _a="odinho"><a href="#l-391">#</a> [21:09] &lt;odinho&gt; Monster merge commit.</li>
 <li id="l-392" _a="odinho" _r="(fake"><a href="#l-392">#</a> [21:09] &lt;odinho&gt; (fake, but yeah)</li>
 <li id="l-393" _a="odinho"><a href="#l-393">#</a> [21:10] &lt;odinho&gt; I don't think we should review it. But can someone else just randomly go through it and click OK on everything? :P</li>
 <li id="l-394" _a="AutomatedTester" _r="odinho"><a href="#l-394">#</a> [21:10] &lt;AutomatedTester&gt; odinho: whats that? you want to review it</li>
 <li id="l-395" _a="AutomatedTester"><a href="#l-395">#</a> [21:11] &lt;AutomatedTester&gt; one sec while I update critic for you</li>
 <li id="l-396" _a="odinho"><a href="#l-396">#</a> [21:11] &lt;odinho&gt; Huh?</li>
 <li id="l-397" _a="AutomatedTester"><a href="#l-397">#</a> [21:11] &lt;AutomatedTester&gt; 20:11 odinho: I don't think we should review it. But can someone else just randomly go through it and click OK on everything? :P</li>
 <li id="l-398" _a="odinho" _r="Yeah"><a href="#l-398">#</a> [21:11] &lt;odinho&gt; Yeah, because it's mostly just fake conflicts.</li>
 <li id="l-399" _a="odinho"><a href="#l-399">#</a> [21:12] &lt;odinho&gt; It'll be like doing the whole review again.</li>
 <li id="l-400" _a="AutomatedTester" _r="odinho"><a href="#l-400">#</a> [21:12] &lt;AutomatedTester&gt; odinho: thats where my, failed, joke came in</li>
 <li id="l-401" _a="odinho"><a href="#l-401">#</a> [21:12] &lt;odinho&gt; Because of philipj's huge whitespace sweep a while ago.</li>
 <li id="l-402" _a="odinho" _r="OK"><a href="#l-402">#</a> [21:12] &lt;odinho&gt; OK, didn't take it :P</li>
 <li id="l-403" _a="AutomatedTester"><a href="#l-403">#</a> [21:13] &lt;AutomatedTester&gt; I promise to limit my jokes even more now</li>
 <li id="l-404" _a="Ms2ger" _r="Oh"><a href="#l-404">#</a> [21:13] &lt;Ms2ger&gt; Oh, it was a joke?</li>
 <li id="l-405"><a href="#l-405">#</a> [21:13] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-406"><a href="#l-406">#</a> [21:17] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-407"><a href="#l-407">#</a> [21:17] -gitbot:#testing- [web-platform-tests] Velmont pushed 1 new commit to Velmont/move_ported_idb_tests: <a href="https://github.com/w3c/web-platform-tests/commit/159257b28909f508c5b6912ba416b8158e3779a5">https://github.com/w3c/web-platform-tests/commit/159257b28909f508c5b6912ba416b8158e3779a5</a></li>
 <li id="l-408"><a href="#l-408">#</a> [21:17] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 159257b Odin Hørthe Omdal: fixup! IndexedDB: Delete the Opera-ported Microsoft IDB tests</li>
 <li id="l-409"><a href="#l-409">#</a> [21:17] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-410"><a href="#l-410">#</a> [21:18] * Joins: mdas (~mdas@public.cloak)</li>
 <li id="l-411"><a href="#l-411">#</a> [21:20] * Quits: mdas (~mdas@public.cloak) (Client closed connection)</li>
 <li id="l-412"><a href="#l-412">#</a> [21:21] * Quits: plh (plehegar@public.cloak) (&quot;Leaving&quot;)</li>
 <li id="l-413"><a href="#l-413">#</a> [21:24] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-414" _a="heycam|away"><a href="#l-414">#</a> [21:24] * heycam|away is now known as heycam</li>
 <li id="l-415"><a href="#l-415">#</a> [21:26] * Joins: gitbot (~gitbot@public.cloak)</li>
 <li id="l-416"><a href="#l-416">#</a> [21:26] -gitbot:#testing- [web-platform-tests] Velmont pushed 1 new commit to Velmont/move_ported_idb_tests: <a href="https://github.com/w3c/web-platform-tests/commit/765c68ffba263f8dd05eda098482ad0ae08568b5">https://github.com/w3c/web-platform-tests/commit/765c68ffba263f8dd05eda098482ad0ae08568b5</a></li>
 <li id="l-417"><a href="#l-417">#</a> [21:26] -gitbot:#testing- web-platform-tests/Velmont/move_ported_idb_tests 765c68f Odin Hørthe Omdal: fixup! IndexedDB: Whitespace cleanup</li>
 <li id="l-418"><a href="#l-418">#</a> [21:26] * Parts: gitbot (~gitbot@public.cloak) (gitbot)</li>
 <li id="l-419"><a href="#l-419">#</a> [21:28] * Quits: Lachy (~Lachy@public.cloak) (&quot;My MacBook Pro has gone to sleep. ZZZzzz…&quot;)</li>
 <li id="l-420"><a href="#l-420">#</a> [21:31] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-421"><a href="#l-421">#</a> [21:32] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-422"><a href="#l-422">#</a> [21:39] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-423"><a href="#l-423">#</a> [21:55] * Joins: dom (dom@public.cloak)</li>
 <li id="l-424"><a href="#l-424">#</a> [22:02] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-425"><a href="#l-425">#</a> [22:02] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-426"><a href="#l-426">#</a> [22:02] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-427"><a href="#l-427">#</a> [22:25] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-428"><a href="#l-428">#</a> [22:27] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
 <li id="l-429"><a href="#l-429">#</a> [22:27] * Joins: zcorpan (~zcorpan@public.cloak)</li>
 <li id="l-430"><a href="#l-430">#</a> [22:30] * Quits: rhauck (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-431"><a href="#l-431">#</a> [22:32] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-432"><a href="#l-432">#</a> [22:33] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-433"><a href="#l-433">#</a> [22:34] * Quits: zcorpan (~zcorpan@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-434"><a href="#l-434">#</a> [22:46] * Quits: Ms2ger (~Ms2ger@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-435"><a href="#l-435">#</a> [22:54] * Joins: Lachy (~Lachy@public.cloak)</li>
 <li id="l-436"><a href="#l-436">#</a> [23:03] * Quits: rhauck (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-437"><a href="#l-437">#</a> [23:03] * Joins: rhauck (~Adium@public.cloak)</li>
 <li id="l-438"><a href="#l-438">#</a> [23:05] * Joins: rhauck1 (~Adium@public.cloak)</li>
 <li id="l-439"><a href="#l-439">#</a> [23:05] * Quits: rhauck (~Adium@public.cloak) (Client closed connection)</li>
 <li id="l-440"><a href="#l-440">#</a> [23:26] * Joins: glenn (~gadams@public.cloak)</li>
 <li id="l-441"><a href="#l-441">#</a> [23:33] * Quits: glenn (~gadams@public.cloak) (Ping timeout: 180 seconds)</li>
 <li id="l-442" _a="heycam"><a href="#l-442">#</a> [23:39] * heycam is now known as heycam|away</li>
 <li id="l-443"><a href="#l-443">#</a> [23:39] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
 <li id="l-444" _a="heycam|away"><a href="#l-444">#</a> [23:43] * heycam|away is now known as heycam</li>
 <li id="l-445"><a href="#l-445">#</a> [23:43] * Quits: lukeis (~Adium@public.cloak) (&quot;Leaving.&quot;)</li>
 <li id="l-446"><a href="#l-446">#</a> Session Close: Fri Mar 28 00:00:00 2014</li>
</ol>
<p id="end">The end :)
